Windows 64bit. Fixes for compile errors and warnings. Project files still to be created.
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@24461 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/ptclib/snmpserv.cxx
@@ -194,7 +194,7 @@ static void EncodeOID(PDUType & pdu, const PINDEX & reqID,
 	   // Build the response list
 		PSNMP_VarBindList & vars = pdu.m_variable_bindings;
 		PINDEX i = 0;
-		vars.SetSize(varlist.size());
+		vars.SetSize((int)varlist.size());
 		PSNMP::BindingList::const_iterator Iter = varlist.begin();
 		while (Iter != varlist.end()) {
 		  vars[i].m_name.SetValue(Iter->first);
